==> 03_final_dataset_creation.ipynb <==
## Creation of final database

It includes the creation of populist scores and information on the membership in governmnet and in the opposition.

# packages

import pandas as pd
import numpy as np
import os
import re


# os.chdir('C:/Users/...')

predicted = pd.read_csv("data/all_batches_predicted.csv")
labeled = pd.read_csv("data/labeled_new.csv")
leaders = pd.read_csv("data/leaders_new.csv")

### 3.1 Changing the structure of the datasets (idexes, missing values, dropping columns and filling the missing values)

labeled['combined_index'] = labeled['index'].astype(str) + '_' + labeled['paragraph_number'].astype(str)

predicted.drop(['predicted_label'], axis=1, inplace=True)
labeled.drop(['people-centrism', 'anti-elitism', 'populism'], axis=1, inplace=True)
labeled.rename(columns={'populism_full':'label_number'}, inplace=True)

full_data=pd.concat([predicted, labeled], ignore_index=True)

full_data['date'] = pd.to_datetime(full_data['date'], format='%d/%m/%Y')

full_data=leaders.merge(full_data, how='outer', on=['name', 'sejm'])

full_data['leader'] = full_data['leader'].fillna(0)

# if a person was a gov_memebr or in gov_lersership at any point in the term, it will have score 1
full_data['new_gov_member'] = full_data.groupby(['name', 'sejm'])['gov_member'].transform(lambda x: x.any())
full_data.loc[full_data['new_gov_member'] == True, 'gov_member'] = 1
full_data= full_data.drop('new_gov_member', axis=1)

# if a person was a minister at any point in the term it will have score 1
full_data['new_minister'] = full_data.groupby(['name', 'sejm'])['minister'].transform(lambda x: x.any())
full_data.loc[full_data['new_minister'] == True, 'minister'] = 1
full_data= full_data.drop('new_minister', axis=1)

# if a person was in a president office at any point in the term it will have score 1
full_data['new_president_office'] = full_data.groupby(['name', 'sejm'])['president_office'].transform(lambda x: x.any())
full_data.loc[full_data['new_president_office'] == True, 'president_office'] = 1
full_data= full_data.drop('new_president_office', axis=1)


full_data

full_data.sort_values(['date'])

full_data['party_club_start'] = full_data['party_club_start'].fillna('none')
full_data['party_club_end'] = full_data['party_club_end'].fillna('none')

full_data = full_data[(full_data['party_club_end'] != 'none') & (full_data['party_club_start'] != 'none')]

### 3.2 Creating the populist scores

#1997 before
start_date = '1997-03-21'
end_date = '1997-09-21'
data_range = full_data.loc[(full_data['date'] >= start_date) & (full_data['date'] <= end_date), :]
avg_score = data_range.groupby(['name', 'sejm'])['label_number'].mean()
df_before_97 = pd.DataFrame(avg_score.reset_index()).rename(columns={'label_number': 'before_97', 'sejm': 'sejm'})
df_before_97=pd.melt(df_before_97, id_vars=['name', 'sejm'], value_vars=['before_97'])


#1997 after
start_date_2 = '1997-09-21'
end_date_2 = '1998-03-21'
data_range_2 = full_data.loc[(full_data['date'] >= start_date_2) & (full_data['date'] <= end_date_2), :]
avg_score_2 = data_range_2.groupby(['name', 'sejm'])['label_number'].mean()
df_after_97 = pd.DataFrame(avg_score_2.reset_index()).rename(columns={'label_number': 'after_97', 'sejm': 'sejm'})
df_after_97=pd.melt(df_after_97, id_vars=['name', 'sejm'], value_vars=['after_97'])

#2001 before
start_date = '2001-03-23'
end_date = '2001-09-23'
data_range = full_data.loc[(full_data['date'] >= start_date) & (full_data['date'] <= end_date), :]
avg_score = data_range.groupby(['name', 'sejm'])['label_number'].mean()
df_before_01 = pd.DataFrame(avg_score.reset_index()).rename(columns={'label_number': 'before_01', 'sejm': 'sejm'})
df_before_01=pd.melt(df_before_01, id_vars=['name', 'sejm'], value_vars=['before_01'])

#2001 after
start_date_2 = '2001-09-23'
end_date_2 = '2002-03-23'
data_range_2 = full_data.loc[(full_data['date'] >= start_date_2) & (full_data['date'] <= end_date_2), :]
avg_score_2 = data_range_2.groupby(['name', 'sejm'])['label_number'].mean()
df_after_01 = pd.DataFrame(avg_score_2.reset_index()).rename(columns={'label_number': 'after_01', 'sejm': 'sejm'})
df_after_01=pd.melt(df_after_01, id_vars=['name', 'sejm'], value_vars=['after_01'])

#2005 before
start_date = '2005-03-25'
end_date = '2005-09-25'
data_range = full_data.loc[(full_data['date'] >= start_date) & (full_data['date'] <= end_date), :]
avg_score = data_range.groupby(['name', 'sejm'])['label_number'].mean()
df_before_05 = pd.DataFrame(avg_score.reset_index()).rename(columns={'label_number': 'before_05', 'sejm': 'sejm'})
df_before_05=pd.melt(df_before_05, id_vars=['name', 'sejm'], value_vars=['before_05'])

#2005 after
start_date_2 = '2005-09-25'
end_date_2 = '2006-03-25'
data_range_2 = full_data.loc[(full_data['date'] >= start_date_2) & (full_data['date'] <= end_date_2), :]
avg_score_2 = data_range_2.groupby(['name', 'sejm'])['label_number'].mean()
df_after_05 = pd.DataFrame(avg_score_2.reset_index()).rename(columns={'label_number': 'after_05', 'sejm': 'sejm'})
df_after_05=pd.melt(df_after_05, id_vars=['name', 'sejm'], value_vars=['after_05'])

#2007 before
start_date = '2007-04-21'
end_date = '2007-10-21'
data_range = full_data.loc[(full_data['date'] >= start_date) & (full_data['date'] <= end_date), :]
avg_score = data_range.groupby(['name', 'sejm'])['label_number'].mean()
df_before_07 = pd.DataFrame(avg_score.reset_index()).rename(columns={'label_number': 'before_07', 'sejm': 'sejm'})
df_before_07=pd.melt(df_before_07, id_vars=['name', 'sejm'], value_vars=['before_07'])

#2007 after
start_date_2 = '2007-10-21'
end_date_2 = '2008-04-21'
data_range_2 = full_data.loc[(full_data['date'] >= start_date_2) & (full_data['date'] <= end_date_2), :]
avg_score_2 = data_range_2.groupby(['name', 'sejm'])['label_number'].mean()
df_after_07 = pd.DataFrame(avg_score_2.reset_index()).rename(columns={'label_number': 'after_07', 'sejm': 'sejm'})
df_after_07=pd.melt(df_after_07, id_vars=['name', 'sejm'], value_vars=['after_07'])



#2011 before
start_date = '2011-04-09'
end_date = '2011-10-09'
data_range = full_data.loc[(full_data['date'] >= start_date) & (full_data['date'] <= end_date), :]
avg_score = data_range.groupby(['name', 'sejm'])['label_number'].mean()
df_before_11 = pd.DataFrame(avg_score.reset_index()).rename(columns={'label_number': 'before_11', 'sejm': 'sejm'})
df_before_11=pd.melt(df_before_11, id_vars=['name', 'sejm'], value_vars=['before_11'])

#2011 after
start_date_2 = '2011-10-09'
end_date_2 = '2012-04-09'
data_range_2 = full_data.loc[(full_data['date'] >= start_date_2) & (full_data['date'] <= end_date_2), :]
avg_score_2 = data_range_2.groupby(['name', 'sejm'])['label_number'].mean()
df_after_11 = pd.DataFrame(avg_score_2.reset_index()).rename(columns={'label_number': 'after_11', 'sejm': 'sejm'})
df_after_11=pd.melt(df_after_11, id_vars=['name', 'sejm'], value_vars=['after_11'])

# 2015 before
start_date = '2015-04-25'
end_date = '2015-10-25'
data_range = full_data.loc[(full_data['date'] >= start_date) & (full_data['date'] <= end_date), :]
avg_score = data_range.groupby(['name', 'sejm'])['label_number'].mean()
df_before_15 = pd.DataFrame(avg_score.reset_index()).rename(columns={'label_number': 'before_15', 'sejm': 'sejm'})
df_before_15=pd.melt(df_before_15, id_vars=['name', 'sejm'], value_vars=['before_15'])

# 2015 after
start_date_2 = '2015-10-25'
end_date_2 = '2016-04-25'
data_range_2 = full_data.loc[(full_data['date'] >= start_date_2) & (full_data['date'] <= end_date_2), :]
avg_score_2 = data_range_2.groupby(['name', 'sejm'])['label_number'].mean()
df_after_15 = pd.DataFrame(avg_score_2.reset_index()).rename(columns={'label_number': 'after_15', 'sejm': 'sejm'})
df_after_15=pd.melt(df_after_15, id_vars=['name', 'sejm'], value_vars=['after_15'])


full_data.drop(['debate', 'topic', 'link','speech', 'session', 'date', 'index', 'paragraph_number', 'combined_index', 'label_number','legislator'], axis=1, inplace=True)
full_data = full_data.drop_duplicates()
full_data= full_data.reset_index(drop=True)

new_df=pd.concat([df_before_97,df_after_97, df_before_01,df_after_01, df_before_07,df_after_07, df_before_11,df_after_11, df_before_15,df_after_15])
merged_df = new_df.merge(full_data, how='left', on=['name', 'sejm'])
merged_df.sort_values(['name','sejm'])

merged_df["index"] = merged_df["variable"].str.replace("(before|after)_\d+", "\\1")
merged_df = merged_df.drop("variable", axis=1)
wide_data = merged_df.pivot(index=["name", "sejm", "leader", "minister","gov_member","president_office","party_club_start","party_club_end","change"], columns="index", values="value")
wide_data = wide_data.reset_index()

wide_data

grouped = merged_df.groupby(['name', 'sejm']).apply(
    lambda x: x.iloc[1]['value'] - x.iloc[0]['value'] if len(x) > 1 else 0
).reset_index(name='diff')

grouped= grouped.reset_index(drop=True)


grouped_df = grouped.merge(full_data, how='left', on=['name', 'sejm'])
grouped_df.sort_values(['name','sejm'])

# diff= BEFORE-AFTER eg. Just before elections in 2011 - just after the elections in 2007

### 3.3 Adding information on the membership in government and in opposition

# information on the membership in govermnet and in opposition
# term 2
gov_2=['SLD', 'PSL']	
opp_2=['UD','UW', 'AWS', 'BBWR', 'KPN', 'PPS']

# term 3
gov_3=['AWS', 'UW']
opp_3=['PSL', 'SLD', 'ROP']	

# term 4
gov_4=['SLD-UP', 'PSL', 'SLD']
opp_4=['PSL', 'PO', 'PIS', 'LPR', 'SKL', 'RKN', 'PP', 'ROP', 'Dom Ojczysty']

# term 5
gov_5=['PIS', 'Samoobrona', 'LPR']
opp_5=['SLD', 'PSL', 'PO']

# term 6
gov_6= ['PO', 'PSL']
opp_6=['SLD', 'PIS', 'SDPL', 'PJN']

# term 7
gov_7=['PO', 'PSL']
opp_7=['SLD', 'PIS', 'ZP']

#term 8
gov_8=['PIS']
opp_8=['SLD', 'PSL-KP', 'PO-KO', 'Teraz', 'Nowoczesna']



grouped_df['gov'] = 0
mask_start = grouped_df['party_club_start'].isin(gov_2) & (grouped_df['sejm'] == 2)
mask_end = grouped_df['party_club_end'].isin(gov_2) & (grouped_df['sejm'] == 2)
grouped_df.loc[mask_start | mask_end, 'gov'] = 1

mask_start = grouped_df['party_club_start'].isin(gov_3) & (grouped_df['sejm'] == 3)
mask_end = grouped_df['party_club_end'].isin(gov_3) & (grouped_df['sejm'] == 3)
grouped_df.loc[mask_start | mask_end, 'gov'] = 1

mask_start = grouped_df['party_club_start'].isin(gov_4) & (grouped_df['sejm'] == 4)
mask_end = grouped_df['party_club_end'].isin(gov_4) & (grouped_df['sejm'] == 4)
grouped_df.loc[mask_start | mask_end, 'gov'] = 1

mask_start = grouped_df['party_club_start'].isin(gov_5) & (grouped_df['sejm'] == 5)
mask_end = grouped_df['party_club_end'].isin(gov_5) & (grouped_df['sejm'] == 5)
grouped_df.loc[mask_start | mask_end, 'gov'] = 1

mask_start = grouped_df['party_club_start'].isin(gov_6) & (grouped_df['sejm'] == 6)
mask_end = grouped_df['party_club_end'].isin(gov_6) & (grouped_df['sejm'] == 6)
grouped_df.loc[mask_start | mask_end, 'gov'] = 1

mask_start = grouped_df['party_club_start'].isin(gov_7) & (grouped_df['sejm'] == 7)
mask_end = grouped_df['party_club_end'].isin(gov_7) & (grouped_df['sejm'] == 7)
grouped_df.loc[mask_start | mask_end, 'gov'] = 1

mask_start = grouped_df['party_club_start'].isin(gov_8) & (grouped_df['sejm'] == 8)
mask_end = grouped_df['party_club_end'].isin(gov_8) & (grouped_df['sejm'] == 8)
grouped_df.loc[mask_start | mask_end, 'gov'] = 1


grouped_df['opp'] = 0
mask_start = grouped_df['party_club_start'].isin(opp_2) & (grouped_df['sejm'] == 2)
mask_end = grouped_df['party_club_end'].isin(opp_2) & (grouped_df['sejm'] == 2)
grouped_df.loc[mask_start | mask_end, 'opp'] = 1

mask_start = grouped_df['party_club_start'].isin(opp_3) & (grouped_df['sejm'] == 3)
mask_end = grouped_df['party_club_end'].isin(opp_3) & (grouped_df['sejm'] == 3)
grouped_df.loc[mask_start | mask_end, 'gov'] = 1

mask_start = grouped_df['party_club_start'].isin(opp_4) & (grouped_df['sejm'] == 4)
mask_end = grouped_df['party_club_end'].isin(opp_4) & (grouped_df['sejm'] == 4)
grouped_df.loc[mask_start | mask_end, 'gov'] = 1

mask_start = grouped_df['party_club_start'].isin(opp_5) & (grouped_df['sejm'] == 5)
mask_end = grouped_df['party_club_end'].isin(opp_5) & (grouped_df['sejm'] == 5)
grouped_df.loc[mask_start | mask_end, 'gov'] = 1

mask_start = grouped_df['party_club_start'].isin(opp_6) & (grouped_df['sejm'] == 6)
mask_end = grouped_df['party_club_end'].isin(opp_6) & (grouped_df['sejm'] == 6)
grouped_df.loc[mask_start | mask_end, 'gov'] = 1

mask_start = grouped_df['party_club_start'].isin(opp_7) & (grouped_df['sejm'] == 7)
mask_end = grouped_df['party_club_end'].isin(opp_7) & (grouped_df['sejm'] == 7)
grouped_df.loc[mask_start | mask_end, 'gov'] = 1

mask_start = grouped_df['party_club_start'].isin(opp_8) & (grouped_df['sejm'] == 8)
mask_end = grouped_df['party_club_end'].isin(opp_8) & (grouped_df['sejm'] == 8)
grouped_df.loc[mask_start | mask_end, 'gov'] = 1

grouped_df

#grouped_df = grouped_df.drop_duplicates()

grouped_df.to_csv('data/scores.csv', encoding='utf-8', index=False)
